#99: MySearch View 분리 위한 코드 리팩토링
